+++ changelog.md
@@ -6,6 +6,9 @@
   - Added [Hide trailing line](docs/preferences.md#display) display preference to hide trailing empty line.
 - **Fixed:**
   - Display:
+    - Calculate split bar movement using current split bar height
+    - Use current scrollbar height/widths instead of fixed #s to allow for better theme support
+    - Correctly factor in current theme display padding into available character width and height
     - Fixed vertical scroll bar when split button shown
     - Fixed show split button being shown after turning off split scroll
     - Readd - and _ as selectable characters for double click and context menu word selection